requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
plotly>=5.20.0
jupyter
pytest
PyYAML>=6.0
